gb/ppu: move LCDStatus struct into register module
